Anpassungen für SPA
