+++ page-details.html
@@ -1,1 +1,1 @@
-//details
+<h1>Test</h1>
